[feat] vegan magazine indicator 추가
