+++ app/build.gradle
@@ -64,4 +64,8 @@ dependencies {
 
     // Splash Screen
     implementation 'androidx.core:core-splashscreen:1.0.0'
+
+    // indicator
+    implementation 'me.relex:circleindicator:2.1.6'
+    implementation 'me.relex:circleindicator:1.3.2'
 }
