+++ app/src/main/java/com/example/beginvegan/src/ui/view/MainActivity.kt
@@ -1,19 +1,20 @@
 package com.example.beginvegan.src.ui.view
 
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
 import com.example.beginvegan.R
 import com.example.beginvegan.config.BaseActivity
 import com.example.beginvegan.databinding.ActivityMainBinding
+import com.example.beginvegan.src.ui.view.mainhome.MainHomeFragment
 
 class MainActivity : BaseActivity<ActivityMainBinding>({ ActivityMainBinding.inflate(it)}) {
     override fun init() {
-        supportFragmentManager.beginTransaction().replace(R.id.fl_main,MainHomeFragment()).commit()
+        supportFragmentManager.beginTransaction().replace(R.id.fl_main, MainHomeFragment()).commit()
 
         binding.bnvMain.setOnItemSelectedListener {
             when(it.itemId){
                 R.id.item_home->{
-                    supportFragmentManager.beginTransaction().replace(R.id.fl_main,MainHomeFragment()).commit()
+                    supportFragmentManager.beginTransaction().replace(R.id.fl_main,
+                        MainHomeFragment()
+                    ).commit()
                 }
                 R.id.item_map->{
                     //supportFragmentManager.beginTransaction().replace(R.id.fl_main,MainHomeFragment()).commit()
@@ -28,4 +29,6 @@ class MainActivity : BaseActivity<ActivityMainBinding>({ ActivityMainBinding.inf
             true
         }
     }
+
+
 }
