--- app/src/main/java/com/example/beginvegan/src/ui/view/MainHomeFragment.kt
@@ -1,29 +1,0 @@
-package com.example.beginvegan.src.ui.view
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.core.view.marginEnd
-import androidx.viewpager2.widget.ViewPager2
-import com.example.beginvegan.R
-import com.example.beginvegan.config.BaseFragment
-import com.example.beginvegan.databinding.FragmentMainHomeBinding
-import com.google.android.material.tabs.TabLayout
-import com.google.android.material.tabs.TabLayoutMediator
-
-class MainHomeFragment : BaseFragment<FragmentMainHomeBinding>(
-    FragmentMainHomeBinding::bind,R.layout.fragment_main_home ){
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        val viewpager = binding.vpMagazines
-        val homeMagazineAdapter = HomeMagazineAdapter(this)
-        viewpager.adapter = homeMagazineAdapter
-
-        TabLayoutMediator(binding.tlHomeMagazineIndicators, viewpager) {tab,position ->}.attach()
-    }
-
-}
