+++ app/src/main/java/com/example/beginvegan/src/ui/view/mainhome/HomeRecommendRestAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.beginvegan.src.ui.view
+package com.example.beginvegan.src.ui.view.mainhome
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
