+++ app/src/main/res/layout/fragment_home_magazine_vegan_define.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".src.ui.view.HomeMagazineVeganDefineFragment"
+    tools:context=".src.ui.view.mainhome.HomeMagazineVeganDefineFragment"
     >
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
