+++ app/src/main/res/layout/fragment_home_magazine_vegan_types.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".src.ui.view.HomeMagazineVeganTypesFragment"
+    tools:context=".src.ui.view.mainhome.HomeMagazineVeganTypesFragment"
     >
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
